+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Axis: Added rules for zoom in/out boundaries, axis span limits, and square ratio locking (#3139, #3142)
 * ErrorBar: Improved axis limit detection for data that does not start at zero (#3155) _Thanks @wolfcomp_
 * DataSources: Created `SignalSourceUInt16` to demonstrate how to plot data with custom types (#3154) _Thanks @angulion_
+* Signal Plot: Added support for generic type arrays and lists (#3154)
 
 ## ScottPlot 5.0.11-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-12-29_